Refactor UserEventsHandler
