+++ src/views/1_ApplicationComponent.tsx
@@ -1,4 +1,6 @@
 import {SessionComponent} from "./2_SessionComponent";
+import {PromptComponent} from "./4_PromptComponent";
+import {JobComponent} from "./3_JobComponent";
 import {TabComponent, TabProps, Tab} from "./TabComponent";
 import * as React from "react";
 import * as _ from "lodash";
@@ -9,6 +11,10 @@ import {saveWindowBounds} from "./ViewUtils";
 import {StatusBarComponent} from "./StatusBarComponent";
 import {PaneTree, Pane} from "../utils/PaneTree";
 import {SearchComponent} from "./SearchComponent";
+import {isKeybindingForEvent} from "./keyevents/Keybindings";
+import {KeyboardAction} from "../Enums";
+import {UserEvent} from "../Interfaces";
+import {isModifierKey} from "./ViewUtils";
 
 export class ApplicationComponent extends React.Component<{}, {}> {
     private tabs: Tab[] = [];
@@ -99,7 +105,6 @@ export class ApplicationComponent extends React.Component<{}, {}> {
         this.focusTab(newPosition + 1);
     }
 
-    // FIXME: this method should be private.
     closeFocusedPane() {
         this.focusedTab.closeFocusedPane();
 
@@ -108,6 +113,198 @@ export class ApplicationComponent extends React.Component<{}, {}> {
         }
 
         this.forceUpdate();
+    }
+
+    handleUserEvent(
+        session: SessionComponent,
+        job: JobComponent,
+        prompt: PromptComponent,
+        search: SearchComponent,
+        event: UserEvent,
+    ) {
+        // Pasted data
+        if (event instanceof ClipboardEvent) {
+            if (search.isFocused) {
+                return;
+            }
+
+            if (!job.props.job.isInProgress()) {
+                prompt.focus();
+                event.preventDefault();
+                prompt.appendText(event.clipboardData.getData("text/plain"));
+                return;
+            }
+
+            job.props.job.write(event.clipboardData.getData("text/plain"));
+
+            event.stopPropagation();
+            event.preventDefault();
+
+            return;
+        }
+
+        // Close focused pane
+        if (isKeybindingForEvent(event, KeyboardAction.paneClose) && !job.props.job.isInProgress()) {
+            this.closeFocusedPane();
+
+            this.forceUpdate();
+
+            event.stopPropagation();
+            event.preventDefault();
+            return;
+        }
+
+        // Change focussed tab
+        if (isKeybindingForEvent(event, KeyboardAction.tabFocus)) {
+            const position = parseInt(event.key, 10);
+            this.focusTab(position);
+
+            event.stopPropagation();
+            event.preventDefault();
+            return;
+        }
+
+        // Enable debug mode
+        if (isKeybindingForEvent(event, KeyboardAction.developerToggleDebugMode)) {
+            window.DEBUG = !window.DEBUG;
+
+            require("devtron").install();
+            console.log(`Debugging mode has been ${window.DEBUG ? "enabled" : "disabled"}.`);
+
+            this.forceUpdate();
+
+            event.stopPropagation();
+            event.preventDefault();
+            return;
+        }
+
+        // Console clear
+        if (isKeybindingForEvent(event, KeyboardAction.cliClearJobs) && !job.props.job.isInProgress()) {
+            session.props.session.clearJobs();
+
+            event.stopPropagation();
+            event.preventDefault();
+            return;
+        }
+
+        if (event.metaKey) {
+            event.stopPropagation();
+            // Don't prevent default to be able to open developer tools and such.
+            return;
+        }
+
+        if (search.isFocused) {
+            // Search close
+            if (isKeybindingForEvent(event, KeyboardAction.editFindClose)) {
+                search.clearSelection();
+                setTimeout(() => prompt.focus(), 0);
+
+                event.stopPropagation();
+                event.preventDefault();
+                return;
+            }
+
+            return;
+        }
+
+        if (job.props.job.isRunningPty() && !isModifierKey(event)) {
+            // CLI interrupt
+            if (isKeybindingForEvent(event, KeyboardAction.cliInterrupt)) {
+                job.props.job.interrupt();
+            } else {
+                job.props.job.write(event);
+            }
+
+            event.stopPropagation();
+            event.preventDefault();
+            return;
+        }
+
+        prompt.focus();
+
+        // Append last argument to prompt
+        if (isKeybindingForEvent(event, KeyboardAction.cliAppendLastArgumentOfPreviousCommand)) {
+            prompt.appendLastLArgumentOfPreviousCommand();
+
+            event.stopPropagation();
+            event.preventDefault();
+            return;
+        }
+
+        if (!job.props.job.isInProgress()) {
+            // CLI Delete word
+            if (isKeybindingForEvent(event, KeyboardAction.cliDeleteWord)) {
+                prompt.deleteWord();
+
+                event.stopPropagation();
+                event.preventDefault();
+                return;
+            }
+
+            // CLI execute command
+            if (isKeybindingForEvent(event, KeyboardAction.cliRunCommand)) {
+                prompt.execute((event.target as HTMLElement).innerText);
+
+                event.stopPropagation();
+                event.preventDefault();
+                return;
+            }
+
+            // CLI clear
+            if (isKeybindingForEvent(event, KeyboardAction.cliClearText)) {
+                prompt.clear();
+
+                event.stopPropagation();
+                event.preventDefault();
+                return;
+            }
+
+            if (prompt.isAutocompleteShown()) {
+                if (isKeybindingForEvent(event, KeyboardAction.autocompleteInsertCompletion)) {
+                    prompt.applySuggestion();
+
+                    event.stopPropagation();
+                    event.preventDefault();
+                    return;
+                }
+
+                if (isKeybindingForEvent(event, KeyboardAction.autocompletePreviousSuggestion)) {
+                    prompt.focusPreviousSuggestion();
+
+                    event.stopPropagation();
+                    event.preventDefault();
+                    return;
+                }
+
+                if (isKeybindingForEvent(event, KeyboardAction.autocompleteNextSuggestion)) {
+                    prompt.focusNextSuggestion();
+
+                    event.stopPropagation();
+                    event.preventDefault();
+                    return;
+                }
+            } else {
+                if (isKeybindingForEvent(event, KeyboardAction.cliHistoryPrevious)) {
+                    prompt.scrollIntoView();
+                    prompt.setPreviousHistoryItem();
+
+                    event.stopPropagation();
+                    event.preventDefault();
+                    return;
+                }
+
+                if (isKeybindingForEvent(event, KeyboardAction.cliHistoryNext)) {
+                    prompt.scrollIntoView();
+                    prompt.setNextHistoryItem();
+
+                    event.stopPropagation();
+                    event.preventDefault();
+                    return;
+                }
+            }
+        }
+
+        prompt.setPreviousKeyCode(event);
     }
 
     render() {
